+++ src/utilities.h
@@ -15,5 +15,6 @@ Napi::Value format(const Napi::CallbackInfo& info);
 void block(const Napi::CallbackInfo& info);
 Napi::Value _maxColourDistance(const Napi::CallbackInfo& info);
 Napi::Value _isUsingJemalloc(const Napi::CallbackInfo& info);
+Napi::Value _isUsingX64V2(const Napi::CallbackInfo& info);
 
 #endif  // SRC_UTILITIES_H_